Do some stuff to automatically make events
We've now got an ObanExample.Business worker that should be able to
manage jobs that have been enqueued. Additionally, a JobMachine which is
a GenServer that calls a function and schedules a job every 60 seconds.

The Business worker should process two jobs:

one inserts an event into the database and also schedule a
job to delete that event 60 seconds later. This is wrapped in a
transaction to demonstrate why running Oban on the same DB as your
application is advantageous.

the other processes the delete job and deletes the event.